Make sure cache memoization works across requests
The $class parameter of cache_memoized() can be an object in some
situations, so the memoize cache would have an entry per instance.

Given that we create component instances on each call, this cache would
grow very fast, and the memoization would not work across multiple
requests or even multiple calls to the same component.

To fix, we now force the $class to be just the class name.

Signed-off-by: Pedro Melo <melo@simplicidade.org>